+++ dart/collision/bullet/BulletTypes.hpp
@@ -33,11 +33,8 @@
 #ifndef DART_COLLISION_BULLET_BULLETTYPES_HPP_
 #define DART_COLLISION_BULLET_BULLETTYPES_HPP_
 
-// Must be included before any Bullet headers.
 #include <Eigen/Dense>
 #include <btBulletCollisionCommon.h>
-
-#include "dart/config.hpp"
 
 namespace dart {
 namespace collision {